Implement investmentType CRUD

+++ src/main/java/com/alipour/product/financialtracker/investment_type/models/InvestmentType.java
@@ -27,7 +27,7 @@ public class InvestmentType extends ParentEntity {
     private String code;
 
     @Column(name = "latest_price")
-    private Float latestPrice;
+    private Double latestPrice;
 
     public InvestmentType(Long id) {
         super(id);
